Take slip in basis points and convert to floats (divide by 10000)

+++ internal/models/eventSwap.go
@@ -14,7 +14,7 @@ type EventSwap struct {
 	Event
 	Pool         common.Asset
 	PriceTarget  int64
-	TradeSlip    float64
+	TradeSlip    int64
 	LiquidityFee int64
 }
